requirements.txt
Faker==0.7.12
pytest==3.6.0
pytest-xdist==1.23.2
allure-pytest==2.5.3
PyYAML==3.13
requests==2.17.3
retrying==1.3.3
urllib3==1.21.1
PyHamcrest==1.9.0